test: Implement github-task script properly
This is a script that chooses a task from GitHub based
on Status API, open pull requests and so on. This functionality
used to be included in 'check-verify --next'

This is split out from check-verify so that it can run other
things besides check-verify

Closes #3335
